+++ test/correctness/PySys_internal_121/run.py
@@ -30,6 +30,7 @@ pysys
 threadPoolMaxWorkers
 isInterruptTerminationInProgress
 isInterruptTerminationInProgressEvent
+isInterruptTerminationInProgressHandle
 grepWarnIfLineLongerThan
 		""".strip().split('\n')